Added Conversion as implementation of Reaction. Found out that Simulation has to be split into subobjects to avoid circular dependencies.

+++ include/Conversion.h
@@ -17,10 +17,10 @@ class Conversion : public Reaction
 			double inBackwardRate);
 		
 		double performForward(
-			std::vector<unsigned long long> particleIndices,
+			std::vector<unsigned long int> particleIndices,
 			Simulation* simulation);
 		double performBackward(
-			std::vector<unsigned long long> particleIndices,
+			std::vector<unsigned long int> particleIndices,
 			Simulation* simulation);
 };
